Add raw deployment deployer (#3075)
* Add kubernetes deployment deployer

* Run update-codegen.sh

* Remove unneeded function

* Address build issues

* Run make schema-generate

* Get logs by selector in integration_test

This allows to get the logs of the deployment too

* Add implementations for Lister, Describer and Remover for raw deployments

* Use correct deploytype in integration test

* tmp: Use http scheme for url in describer

* Fix receiving pod logs correctly in test

* Fix deployer integration test

* Wait for pods of deployment to be ready too

* Move Deployer, Describer, Lister and Remover int tests to correct package

* Log receiver also check existing pods

* Revert "Log receiver also check existing pods"

This reverts commit a9bf27447fcf53ae285d720d1d9d4a2f13d45f7d.

* tmp: disable cleanup to get pod logs / status and add more logs

* Revert "tmp: disable cleanup to get pod logs / status and add more logs"

This reverts commit b4987242cf22ef7f3bc9e6014bbb0574b39efd2e.

* Give a bit more time to collect logs

* Add multi deployer, -lister, -remover

* Move deployer, lister, describer and removers in separate packages

* Use integration tests in k8s packages too

* Fix import cycle issue

* Fix integration tests

* Remove unneeded test

* Run all deployer integration tests on both deployer implementations

* Add small delay in test

* Revert "Add small delay in test"

This reverts commit 2aec3f4146cae4278e2a5042e59a691646d9e36a.

* Wait for deployments to be up

* wait for ready pods instead of checking only if containers are running

* tmp: collect pod status and logs

* Revert "tmp: collect pod status and logs"

This reverts commit 8bc2a0bb437ca6100f2f0fae3843884523425c05.

* Add deploy-type information in description

* Poll for pod logs instead of time.sleep

* Switch to PollUntilContextTimeout in wait functions

* Make sure no pods of old replicas of the deployment are running anymore in WaitForDeploymentsAvailable functions

* Change clients WithLister|Describer|Remover methods to variadic arguments

* Move Deployers, Listers, Describers and Removers back into knative/k8s package

* [WIP] Resolve some cyclic dependencies

* Remove subscribe functionality from raw deployer to remove dependencies

* Some more cleanup and import alias fixes

* Rename deploy-type arg to deployer to align with other arg names

* Fix build issues in cmd unit tests

* Adjust integration tests to not test func subscribe by default and skip test for kubernetes deployer

* Add shell completion for deployer selection

* Undo unneeded changes in imports ordering

* Move SynchronizedBuffer to k8s package

* Readd Test_processValue unit test

* Use constants for waiting timeout

* Use onClusterFix for raw deployer too

* Remove unneeded build tags

* Use TestInt_ pattern for integration tests (and helper functions)

* Use dot import for pkg/testing

* Use dot import for pkg/testing/k8s

* Use fn alias for pkg/functions import for consistency

* Update outdated comment

* Some more TestInt_xyz replaces

* Mark test helper functions with t.Helper()

* Removed build tags in other int_test_helpers.go

* Use dot imports for some testing packages

* Move testing/k8s to testing package

* Revert "Move testing/k8s to testing package"

This reverts commit 9782f418ac5d257bda1b6bf5aca7b9c5382935d1.

* Update describe to return a ErrNotHandled error

* Update remover to return a ErrNotHandled error

* Update lister to only return resultlist and error (remove handled/bool return value)

* Move IsCRDNotFoundError() to knative package and make specific to Knative

* Add +build integration tag on integration test files

* Remove misleading comment

* Fix build issues

* Add a few more comments to the Describe, Remove and List usage about responsibility

* Remove outdated package comment

* Remove need to specify deployer in function programmatic declaration

* create Trigger in test only for Knative Service as we don't support it yet for the raw deployer anyhow

* Use correct API version for triggers subscriber ref

* Ignore golang-lint dot import errors in testing helpers

+++ pkg/knative/remover.go
@@ -22,7 +22,7 @@ type Remover struct {
 	verbose bool
 }
 
-func (remover *Remover) Remove(ctx context.Context, name, ns string) (err error) {
+func (remover *Remover) Remove(ctx context.Context, name, ns string) error {
 	if ns == "" {
 		fmt.Fprintf(os.Stderr, "no namespace defined when trying to delete a function in knative remover\n")
 		return fn.ErrNamespaceRequired
@@ -30,16 +30,35 @@ func (remover *Remover) Remove(ctx context.Context, name, ns string) (err error)
 
 	client, err := NewServingClient(ns)
 	if err != nil {
-		return
+		return err
 	}
+
+	ksvc, err := client.GetService(ctx, name)
+	if err != nil {
+		// If we can't get the service, check why
+		if IsCRDNotFoundError(err) {
+			// Knative Serving not installed - we don't handle this
+			return fn.ErrNotHandled
+		}
+		if apiErrors.IsNotFound(err) {
+			// Service doesn't exist as a Knative service - we don't handle this
+			return fn.ErrNotHandled
+		}
+		// Some other error (permissions, network, etc.) - this is a real error
+		// We can't determine if we should handle it, so propagate it
+		return fmt.Errorf("failed to get knative service: %w", err)
+	}
+
+	if !UsesKnativeDeployer(ksvc.Annotations) {
+		return fn.ErrNotHandled
+	}
+
+	// We're responsible, for this function --> proceed...
 
 	err = client.DeleteService(ctx, name, RemoveTimeout)
 	if err != nil {
-		if apiErrors.IsNotFound(err) {
-			return fn.ErrFunctionNotFound
-		}
-		err = fmt.Errorf("knative remover failed to delete the service: %v", err)
+		return fmt.Errorf("knative remover failed to delete the service: %v", err)
 	}
 
-	return
+	return nil
 }
